+++ packages/cli/src/compile/compile.ts
@@ -8,8 +8,6 @@ import type { CompileOptions } from './command';
 export default async (opts: CompileOptions, log: Logger) => {
   log.debug('Loading job...');
   const compilerOptions: Options = await loadTransformOptions(opts, log);
-  compilerOptions.logger = createLogger(COMPILER, opts as any); // TODO log options are a bit flaky right now
-
   const job = compile(opts.jobSource || opts.jobPath, compilerOptions);
   if (opts.jobPath) {
     log.success(`Compiled job from ${opts.jobPath}`);
@@ -57,25 +55,21 @@ export const loadTransformOptions = async (
   log: Logger
 ) => {
   const options: Options = {
-    logger: log,
+    logger: log || createLogger(COMPILER, opts as any),
   };
-
   // If an adaptor is passed in, we need to look up its declared exports
   // and pass them along to the compiler
-  if (opts.adaptors?.length) {
+  if (opts.adaptors?.length && opts.ignoreImports != true) {
     let exports;
-    const [pattern] = opts.adaptors; // TODO add-imports only takes on adaptor, but the cli can take multiple
+    const [pattern] = opts.adaptors;
     const [specifier] = pattern.split('=');
 
     // Preload exports from a path, optionally logging errors in case of a failure
-    log.debug(`Attempting to preload typedefs for ${specifier}`);
+    log.debug(`Attempting to preload types for ${specifier}`);
     const path = await resolveSpecifierPath(pattern, opts.repoDir, log);
     if (path) {
       try {
-        exports = await preloadAdaptorExports(path);
-        if (exports) {
-          log.info(`Loaded typedefs for ${specifier}`);
-        }
+        exports = await preloadAdaptorExports(path, log);
       } catch (e) {
         log.error(`Failed to load adaptor typedefs from path ${path}`);
         log.error(e);
@@ -87,6 +81,7 @@ export const loadTransformOptions = async (
     }
 
     options['add-imports'] = {
+      ignore: opts.ignoreImports as string[],
       adaptor: {
         name: stripVersionSpecifier(specifier),
         exports,
@@ -94,5 +89,6 @@ export const loadTransformOptions = async (
       },
     };
   }
+
   return options;
 };