+++ packages/runtime-manager/CHANGELOG.md
@@ -1,5 +1,23 @@
 # runtime-manager
 
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies [f6d22ec]
+- Updated dependencies [77bed12]
+- Updated dependencies [8dfc5bf]
+- Updated dependencies [f4b9702]
+  - @openfn/compiler@0.0.28
+  - @openfn/logger@0.0.12
+  - @openfn/runtime@0.0.21
+
+## 0.0.28
+
+### Patch Changes
+
+- @openfn/compiler@0.0.27
+
 ## 0.0.27
 
 ### Patch Changes